+++ app/models/bank.rb
@@ -1,4 +1,6 @@
 # Represents guild banks
 class Bank < ApplicationRecord
   belongs_to :guild
+
+  has_many :bank_cells, dependent: :destroy
 end