+++ app/models/bank_cell.rb
@@ -1,0 +1,4 @@
+# Represents bank cells
+class BankCell < ApplicationRecord
+  belongs_to :bank
+end