+++ spec/factories/bank_cells.rb
@@ -1,0 +1,7 @@
+FactoryBot.define do
+  factory :bank_cell do
+    item_uid { 11_370 }
+    amount { 10 }
+    association :bank
+  end
+end